LPS-72263 move packages
