+++ modules/apps/portal-mobile-device-detection-fiftyonedegrees-lite/portal-mobile-device-recognition-provider-fiftyonedegrees/src/main/java/com/liferay/portal/mobile/device/detection/fiftyonedegrees/internal/FiftyOneDegreesDevice.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.mobile.device.recognition.provider.fiftyonedegrees.internal;
+package com.liferay.portal.mobile.device.detection.fiftyonedegrees.internal;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -21,7 +21,7 @@ import com.liferay.portal.kernel.mobile.device.Capability;
 import com.liferay.portal.kernel.mobile.device.Dimensions;
 import com.liferay.portal.kernel.mobile.device.VersionableName;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.mobile.device.recognition.provider.fiftyonedegrees.internal.constants.FiftyOneDegreesPropertyNames;
+import com.liferay.portal.mobile.device.detection.fiftyonedegrees.internal.constants.FiftyOneDegreesPropertyNames;
 
 import fiftyone.mobile.detection.Match;
 import fiftyone.mobile.detection.entities.Values;
